Role Manager folder, move active chatter role processor.
