+++ backend/variables/builtin/active-chat-user-count.js
@@ -1,6 +1,6 @@
 "use strict";
 const logger = require("../../logwrapper");
-const activeViewerHandler = require('../../chat/active-chatters');
+const activeViewerHandler = require('../../roles/role-managers/active-chatters');
 
 const { OutputDataType } = require("../../../shared/variable-contants");
