+++ shared/firebot-roles.js
@@ -4,6 +4,10 @@ const firebotRoles = [
     {
         id: "ActiveChatters",
         name: "Active Chatters"
+    },
+    {
+        id: "ActiveMixplayUsers",
+        name: "Active Mixplay Users"
     }
 ];
